tests/wpt/tests/cookies/attributes/domain.sub.html

<!doctype html>
<html>
  <head>
    <meta charset=utf-8>
    <title>Test cookie domain attribute parsing</title>
    <meta name=help href="https://tools.ietf.org/html/rfc6265#section-5.2.3">
    <meta name="timeout" content="long">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
  </head>
  <body>
    <script>
      const port    = "{{ports[http][0]}}";
      const wwwHost = "{{domains[www]}}";

      test(t => {
        const win = window.open(`http://${wwwHost}:${port}/cookies/attributes/resources/domain-child.sub.html`);
        fetch_tests_from_window(win);
      });
    </script>
  </body>
</html>
