LayoutTests/imported/w3c/web-platform-tests/cookies/attributes/domain.sub-expected.txt

CONSOLE MESSAGE: Domain is malformed!

Harness Error (FAIL), message = Domain is malformed!

PASS Test cookie domain attribute parsing
